Resolving Controller Button Conflict
Remapped the Driver "return to default position" button from "X" to "B" because there are no manipulator commands mapped to "B" but there is one mapped to "X" and this would be easier to remember. This completes the first iteration of the testing code for being able to Operate Sonic using only one controller.